requirements.txt
pandas
altair
openpyxl
numpy
Pillow
streamlit
streamlit-option-menu
geopandas
matplotlib

